Move AbstractPageParser into Contracts
