+++ src/Contracts/AbstractPageParser.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Hyde\Framework;
+namespace Hyde\Framework\Contracts;
 
 use Hyde\Framework\Actions\ValidatesExistence;
-use Hyde\Framework\Contracts\AbstractPage;
+use Hyde\Framework\PageParserContract;
 
 /**
  * Abstract base class for all page parsers.
